Debug app
There was some mild coupling happening with the joints.

To help with debugging, I included the individual euler angles
for each of the segments.

The final resolution was to just get the euler angles for each of the
the segments and calculate the angle difference without finding
the relative quaternion. I'll investigate this problem later.